fix wonky reporting when two files are added with the same filename
Previously, we would see messages like:

    file has been added after file gathering phase: 'lib/Foo/Bar.pm'

because we were not always saving all file objects - one of the files was
being blown away by the other because it was stored in a hash, indexed by the
common filename.

In this situation, we don't need to report anything extra that we weren't
already checking for -- either one of the files is renamed or duplicated later
(either at the right phase or not), or not (in which case core will report a
duplicate file).